feat: add Raw Response Metadata Handling - GPT Token Usage Example (#270)
Currently it is not really possible to get into details of the
interaction between user and model because the HTTPResponse is hidden
behind the response converter and is then forgotten forever. So the only
possibility currently to get details from the interaction would be a
custom response and a custom response converter.

As a draft and to collect your ideas and other input i want to throw
this stone into the lake with having the HTTPResponse also transported
into the direction of the output processors. Could this give a bit more
flexibility? What do you think?

Please have in mind this is only a draft to discuss. I thought about
having this also possible if a request fails with an error status code
in the future so the output processors could handle in a separation of
convern way if there should be an exception thrown. Or do we maybe need
an additional layer on top of the response converters itself?

For sure this is the simplest use case with the text response in mind.
Maybe for the streamed response it would need a more "recording"
approach with an output processor. But would surely be possible.

---------

Co-authored-by: Christopher Hertel <mail@christopher-hertel.de>