feat: add basic setup for memory injections to system prompt (#387)
This PR introduces a flexible memory system that allows the LLM to
recall contextual information that are permantent for the conversation.
In difference to tools it can be always utilized, when `use_memory`
option is not disabled. It would be possible to fetch memory by a tool
with a system instruction like `always call the tool foo_bar` but, for
me, this feels like a bad design to always force the model to do a tool
call without further need.

Currently i have added just two memory providers to show what my idea
is. I could also think about a write layer to fill a memory, together
with a read layer, this could, for example, be working good with a graph
database and tools for memory handling. But these are ideas for the
future.

So for this first throw i hope you get what i was thinking about to
reach. I decided to inject the memory to the system prompt, when it is
available, instead of adding a second system prompt to the message bag.
But this was just a 50/50 thinking. I tried both seem to be working
equally, at least for open ai. I am open to change it back again.

What do you think?

---------

Co-authored-by: Christopher Hertel <mail@christopher-hertel.de>